requirements.txt
opencv-python
numpy
tensorflow
pandas
python-dotenv
openai
langchain
langchain-openai